# [10.x] Add dependency resolver attributes