--- a/src/Illuminate/Contracts/Container/DependencyResolver.php
+++ b/src/Illuminate/Contracts/Container/DependencyResolver.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Illuminate\Contracts\Container;
+
+interface DependencyResolver
+{
+    /**
+     * Resolve the dependency from the container.
+     *
+     * @param  \Illuminate\Contracts\Container\Container  $container
+     * @return mixed
+     */
+    public function resolve(Container $container): mixed;
+}